requirements.txt
python-telegram-bot==13.15
Flask==2.2.2
gunicorn==20.1.0
aiohttp>=3.8.3,<4.0
Werkzeug==2.2.2  
